+++ src/Craftliltplugin.php
@@ -28,6 +28,7 @@ use lilthq\craftliltplugin\parameters\CraftliltpluginParameters;
 use lilthq\craftliltplugin\records\SettingRecord;
 use lilthq\craftliltplugin\services\appliers\ElementTranslatableContentApplier;
 use lilthq\craftliltplugin\services\appliers\field\FieldContentApplier;
+use lilthq\craftliltplugin\services\handlers\CopySourceTextHandler;
 use lilthq\craftliltplugin\services\handlers\CreateDraftHandler;
 use lilthq\craftliltplugin\services\handlers\CreateJobHandler;
 use lilthq\craftliltplugin\services\handlers\CreateTranslationsHandler;
@@ -43,7 +44,7 @@ use lilthq\craftliltplugin\services\mappers\LanguageMapper;
 use lilthq\craftliltplugin\services\providers\ConnectorConfigurationProvider;
 use lilthq\craftliltplugin\services\providers\ElementTranslatableContentProvider;
 use lilthq\craftliltplugin\services\providers\field\FieldContentProvider;
-use lilthq\craftliltplugin\services\repositories\external\ConnectorJobFileRepository;
+use lilthq\craftliltplugin\services\repositories\external\ConnectorFileRepository;
 use lilthq\craftliltplugin\services\repositories\external\ConnectorJobRepository;
 use lilthq\craftliltplugin\services\repositories\external\ConnectorTranslationRepository;
 use lilthq\craftliltplugin\services\repositories\I18NRepository;
@@ -73,7 +74,7 @@ use yii\web\Response;
  *
  * @property ConnectorJobRepository $connectorJobRepository
  * @property ConnectorTranslationRepository $connectorTranslationRepository
- * @property ConnectorJobFileRepository $connectorJobsFileRepository
+ * @property ConnectorFileRepository $connectorJobsFileRepository
  * @property JobRepository $jobRepository
  * @property TranslationRepository $translationRepository
  * @property ConnectorConfigurationProvider $connectorConfigurationProvider
@@ -99,6 +100,7 @@ use yii\web\Response;
  * @property CreateTranslationsHandler $createTranslationsHandler
  * @property RefreshJobStatusHandler $refreshJobStatusHandler
  * @property CreateDraftHandler $createDraftHandler
+ * @property CopySourceTextHandler $copySourceTextHandler
  * @property ServiceInitializer $serviceInitializer
  */
 class Craftliltplugin extends Plugin