+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.4.0 - 2022-08-03
+### Added
+- Remove entry translation drafts on job removal
+- Empty content now excluded from translation source body
+- Warning for job entries, when translation in progress (on job creation) 
+- Fixed: Empty I18N entry creation
+- Fixed: Translation preview issue (error while encoding of json)
+- Fixed: Hide warning when element index not exist
+- Fixed: Argument 2 passed to craft\services\Drafts::createDraft() must be of the type int, null given
+
+## 0.3.1 - 2022-08-03
+### Added
+- Updated connector sdk dependency 
+
 ## 0.3.0 - 2022-08-03
 ### Added
 - Fixed applying translations for nested fields